requirements.txt
fastapi==0.104.1
uvicorn==0.20.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
jinja2==3.1.2